fix(hash): fix flaky hash computation when multiple parents
The hash computation was flaky when an image has multiple parent
in differents subtrees of the dag. We need to ensure the hash of every
parent has been computed before trying to compute the hash of the child node.
The implementation of the fix will detect the situation and do a repass on
every node until all the parents have a hash computed.