clean(dag): cleanup unused Filter function
Following 7ef5705 the "Filter" function is not used anymore.

After thinking I think it will not ever be used in this form as it is
not generic and thus it is better to completely delete it.

Indeed the function is buggy, difficult to understand (and to fix). E.g. I
still have not grasped what it try to do with the orphans and their
childs.

I think (without being sure) the idea behind this was to be able to work
on a smaller graph for performance but as the small graph is used just
once, the cost of constructing it in the first place completely offsets
the potential gains.